+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - add `as_matching_blocks` to `Editops`/`Opcodes`
 - add support for deletions from `Editops`
 - add `Editops.apply`/`Opcodes.apply`
+- add `Editops.remove_subsequence`
 
 #### Changed
 - merge adjacent similar blocks in `Opcodes`